requirements.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.7.0
boto3==1.28.84
botocore==1.31.84
celery==5.3.5
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.3
MarkupSafe==2.1.3
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.40
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
Werkzeug==3.0.1
